Split up tab views into separate files
