+++ src/client/gui/bb_tab.rs
@@ -1,2 +1,7 @@
 pub mod tab;
 pub mod login;
+mod workout;
+mod social;
+mod mascot;
+mod settings;
+mod home;
