+++ src/client/gui/bb_tab/settings.rs
@@ -1,0 +1,10 @@
+use iced::Element;
+use iced::widget::Row;
+use crate::{Message, UserInterface};
+
+impl UserInterface {
+
+    pub fn settings_screen(&self) -> Element<Message> {
+        Row::new().into()
+    }
+}
