+++ src/client/gui/bb_tab/social.rs
@@ -1,0 +1,10 @@
+use iced::Element;
+use iced::widget::Row;
+use crate::{Message, UserInterface};
+
+impl UserInterface {
+
+    pub fn social_screen(&self) -> Element<Message> {
+        Row::new().into()
+    }
+}
