+++ src/client/gui/bb_tab/tab.rs
@@ -3,9 +3,9 @@ use strum_macros::{Display, EnumIter};
 #[derive(Debug, Clone, Copy, PartialEq, Display, EnumIter)]
 pub enum Tab {
     Home,
-    Workouts,
-    Friends,
-    Mascots,
+    Workout,
+    Social,
+    Mascot,
     Settings,
     Exit
 }
